+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 - **Added:** Support for older daily quests (Ixal and Moogle), which did not return item ids on their objectives.
 - **Changed:** Restricted the tweak to daily quests. Just to make sure it doesn't do anything weird.
-- **Changed:** If the player is not on a crafting job, the recipe search will now open.
+- **Changed:** Just in case if the player is not on a crafter job, the recipe search will now open. Technically, you shouldn't be able to talk to the NPC/progress the quest in the first place, if you're not a crafter.
 - **Fixed:** The check if all materials are available (in your inventory) will now include HQ materials and be calculated correctly when the objective requires you to craft more than one of an item.
 
 ### Character Class Switcher